+++ CHANGELOG.md
@@ -6,7 +6,12 @@ All notable changes to this project will be documented in this file.
 
 What's changed
 
-* 
+* Refactored CLI registry commands to remove some duplication. Resolving the registry with policy checks is common for `generate`, `resolve` and `check`. ([#536](https://github.com/open-telemetry/weaver/pull/536) by @jerbly).
+  * Added missing `after_resolution` policy checks to `generate` and `resolve` through the common code.
+  * Removed the deprecated `--registry-git-sub-dir` option.
+  * Fixed bug in `check` if `--skip-policies` was specified then it would not fail for any validation errors.
+* Semantic Conventions Issue [#1513](https://github.com/open-telemetry/semantic-conventions/issues/1513) - Make span_kind required in yaml and break down multi-kind span definitions - ([#542](https://github.com/open-telemetry/weaver/pull/542) by @jerbly).
+  * Updated the EBNF and JSON schema to define `span_kind` as mandatory for `span` group types. Added a group validity check as a warning.
 
 ## [0.12.0] - 2024-12-09